+++ Frends.AzureBlobStorage.DeleteBlob/Frends.AzureBlobStorage.DeleteBlob.Tests/UnitTests.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using Frends.AzureBlobStorage.DeleteBlob.Definitions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Frends.AzureBlobStorage.DeleteBlob.Tests;
@@ -12,11 +13,16 @@ namespace Frends.AzureBlobStorage.DeleteBlob.Tests;
 [TestClass]
 public class DeleteTest
 {
-    private readonly string _connectionString = Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_ConnString");
+    private readonly string _connectionString =
+        Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_ConnString");
+
     private readonly string _appID = Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_AppID");
     private readonly string _clientSecret = Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_ClientSecret");
     private readonly string _tenantID = Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_TenantID");
-    private readonly string _storageAccount = Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_StorageAccount");
+
+    private readonly string _storageAccount =
+        Environment.GetEnvironmentVariable("Frends_AzureBlobStorage_StorageAccount");
+
     private readonly string _testFileDir = Path.Combine(Environment.CurrentDirectory, "TestFiles");
     private readonly string _firstTestFile = Path.Combine(Environment.CurrentDirectory, "TestFiles", "testfile.txt");
     private readonly string _secondTestFile = Path.Combine(Environment.CurrentDirectory, "TestFiles", "testfile2.txt");
@@ -53,16 +59,12 @@ public class DeleteTest
     {
         var input = new Input
         {
-            BlobName = Guid.NewGuid().ToString(),
-            ContainerName = "none",
-            ConnectionString = _connectionString,
+            BlobName = Guid.NewGuid().ToString(), ContainerName = "none", ConnectionString = _connectionString,
         };
 
         var options = new Options()
         {
-            SnapshotDeleteOption = default,
-            VerifyETagWhenDeleting = default,
-            ThrowErrorIfBlobDoesNotExists = false,
+            SnapshotDeleteOption = default, VerifyETagWhenDeleting = default, ThrowErrorIfBlobDoesNotExists = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, options, default);
@@ -75,16 +77,12 @@ public class DeleteTest
     {
         var input = new Input
         {
-            BlobName = "none",
-            ContainerName = _containerName,
-            ConnectionString = _connectionString,
+            BlobName = "none", ContainerName = _containerName, ConnectionString = _connectionString,
         };
 
         var options = new Options()
         {
-            SnapshotDeleteOption = default,
-            VerifyETagWhenDeleting = default,
-            ThrowErrorIfBlobDoesNotExists = false,
+            SnapshotDeleteOption = default, VerifyETagWhenDeleting = default, ThrowErrorIfBlobDoesNotExists = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, options, default);
@@ -108,9 +106,7 @@ public class DeleteTest
 
         var options = new Options()
         {
-            SnapshotDeleteOption = default,
-            VerifyETagWhenDeleting = default,
-            ThrowErrorIfBlobDoesNotExists = false,
+            SnapshotDeleteOption = default, VerifyETagWhenDeleting = default, ThrowErrorIfBlobDoesNotExists = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, options, default);
@@ -131,9 +127,7 @@ public class DeleteTest
 
         var options = new Options()
         {
-            SnapshotDeleteOption = default,
-            VerifyETagWhenDeleting = default,
-            ThrowErrorIfBlobDoesNotExists = true,
+            SnapshotDeleteOption = default, VerifyETagWhenDeleting = default, ThrowErrorIfBlobDoesNotExists = true,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, options, default);
@@ -195,6 +189,7 @@ public class DeleteTest
     private static BlobContainerClient GetBlobContainer(string connectionString, string containerName)
     {
         var blobServiceClient = new BlobServiceClient(connectionString);
+
         return blobServiceClient.GetBlobContainerClient(containerName);
     }
 
@@ -203,9 +198,13 @@ public class DeleteTest
         Directory.CreateDirectory(_testFileDir);
 
         #region 512 byte files
-        await File.WriteAllTextAsync(_firstTestFile, "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque non sem quis orci rutrum hendrerit. Fusce ultricies cursus ante nec bibendum. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam libero massa, viverra id suscipit in, tincidunt sit amet urna. Vestibulum gravida a massa eu molestie. Phasellus volutpat neque vitae enim molestie, vitae pharetra massa varius. Phasellus ante nulla, faucibus nec tristique eu, dignissim quis magna. Sed vitae sodales ipsum. Ut et maximus nibh. Etiam dui.");
 
-        await File.WriteAllTextAsync(_secondTestFile, "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla mollis neque nibh, molestie consequat lacus euismod in. Phasellus in libero id velit sollicitudin rhoncus. Sed a nunc non lacus hendrerit iaculis. Suspendisse quis dui id enim sollicitudin rhoncus. Phasellus convallis lacinia finibus. Sed quis purus vitae felis finibus facilisis at quis nisi. Integer pharetra, ex egestas iaculis ultricies, tortor neque hendrerit justo, eu vulputate odio eros sed augue. Mauris quis sapien non ligula maximus eget.");
+        await File.WriteAllTextAsync(_firstTestFile,
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque non sem quis orci rutrum hendrerit. Fusce ultricies cursus ante nec bibendum. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam libero massa, viverra id suscipit in, tincidunt sit amet urna. Vestibulum gravida a massa eu molestie. Phasellus volutpat neque vitae enim molestie, vitae pharetra massa varius. Phasellus ante nulla, faucibus nec tristique eu, dignissim quis magna. Sed vitae sodales ipsum. Ut et maximus nibh. Etiam dui.");
+
+        await File.WriteAllTextAsync(_secondTestFile,
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla mollis neque nibh, molestie consequat lacus euismod in. Phasellus in libero id velit sollicitudin rhoncus. Sed a nunc non lacus hendrerit iaculis. Suspendisse quis dui id enim sollicitudin rhoncus. Phasellus convallis lacinia finibus. Sed quis purus vitae felis finibus facilisis at quis nisi. Integer pharetra, ex egestas iaculis ultricies, tortor neque hendrerit justo, eu vulputate odio eros sed augue. Mauris quis sapien non ligula maximus eget.");
+
         #endregion 512 byte files
     }