+++ Frends.AzureBlobStorage.DeleteBlob/Frends.AzureBlobStorage.DeleteBlob/Definitions/Enums.cs
@@ -28,6 +28,8 @@ public enum ConnectionMethod
 {
 #pragma warning disable CS1591 // self explanatory
     ConnectionString,
-    OAuth2
+    OAuth2,
+    ArcManagedIdentity,
+    ArcManagedIdentityCrossTenant,
 #pragma warning restore CS1591 // self explanatory
 }