+++ Frends.AzureBlobStorage.DownloadBlob/Frends.AzureBlobStorage.DownloadBlob/Definitions/Source.cs
@@ -1,5 +1,7 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+
 namespace Frends.AzureBlobStorage.DownloadBlob.Definitions;
 
 /// <summary>
@@ -61,13 +63,34 @@ public class Source
     public string ClientSecret { get; set; }
 
     /// <summary>
-    /// A shared access signature to use when connecting to Azure storage container. 
+    /// A shared access signature to use when connecting to Azure storage container.
     /// Grants restricted access rights to Azure Storage resources when combined with URI.
     /// </summary>
     /// <example>sv=2021-04-10&amp;se=2022-04-10T10%3A431Z&amp;sr=c&amp;sp=l&amp;sig=ZJg983RovE%2BZXI</example>
     [UIHint(nameof(ConnectionMethod), "", ConnectionMethod.SASToken)]
     [PasswordPropertyText]
     public string SASToken { get; set; }
+
+    /// <summary>
+    /// Scopes used when authenticating with Arc Managed Identity Cross Tenant.
+    /// </summary>
+    /// <example>[api://AzureADTokenExchange/.default]</example>
+    [UIHint(nameof(ConnectionMethod), "", ConnectionMethod.ArcManagedIdentityCrossTenant)]
+    public string[] Scopes { get; set; } = Array.Empty<string>();
+
+    /// <summary>
+    /// Target Tenant ID of Azure Tenant.
+    /// </summary>
+    /// <example>Y6b1hf2a-80e2-xyz2-qwer3h-3a7c3a8as4b7f</example>
+    [UIHint(nameof(ConnectionMethod), "", ConnectionMethod.ArcManagedIdentityCrossTenant)]
+    public string TargetTenantId { get; set; }
+
+    /// <summary>
+    /// Target Client ID of Azure Tenant.
+    /// </summary>
+    /// <example>Y6b1hf2a-80e2-xyz2-qwer3h-3a7c3a8as4b7f</example>
+    [UIHint(nameof(ConnectionMethod), "", ConnectionMethod.ArcManagedIdentityCrossTenant)]
+    public string TargetClientId { get; set; }
 
     /// <summary>
     /// Name of the blob to download.
@@ -77,7 +100,7 @@ public class Source
     public string BlobName { get; set; }
 
     /// <summary>
-    /// Set desired content-encoding. 
+    /// Set desired content-encoding.
     /// Defaults to UTF8 BOM.
     /// </summary>
     /// <example>utf8</example>