+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer/Definitions/Enums.cs
@@ -18,7 +18,17 @@ public enum AuthenticationMethod
     /// <summary>
     /// OAuth2.
     /// </summary>
-    OAuth2
+    OAuth2,
+
+    /// <summary>
+    /// Managed Identity.
+    /// </summary>
+    ArcManagedIdentity,
+
+    /// <summary>
+    /// Managed Identity for cross-tenant.
+    /// </summary>
+    ArcManagedIdentityCrossTenant,
 }
 
 /// <summary>