+++ Frends.AzureBlobStorage.CreateContainer/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## [2.2.0] - 2026-01-23
+
+### Added
+
+- Add options to support Arc Managed Identity authentication.
+
 ## [2.1.0] - 2026-01-15
 
 ### Changed
@@ -11,11 +17,13 @@
 ## [2.0.0] - 2025-10-23
 
 ### Added
+
 - Created new **Connection** tab to organize authentication and connection details.
 - Created new **Options** tab with `ThrowErrorOnFailure` and `ErrorMessageOnFailure` fields.
 - Added structured **Error** object (`Message`, `AdditionalInfo`) to task result.
 
 ### Changed
+
 - Moved connection-related parameters from **Input** to **Connection** tab.
 - Renamed:
   - `Input.ConnectionMethod` → `Connection.AuthenticationMethod`
@@ -23,33 +31,46 @@
   - `Input.TenantID` → `Connection.TenantId`
 
 ### Breaking changes
+
 - Parameter moves and renames make this a **breaking change**.
 - To upgrade:
   - Move all connection fields to the **Connection** tab.
 
 ## [1.3.0] - 2024-08-21
+
 ### Updated
+
 - Updated Azure.Identity to version 1.12.0.
 
 ## [1.2.0] - 2024-01-31
+
 ### Updated
+
 - Azure.Identity to version 1.10.4
 - Azure.Storage.Blobs to version 12.19.1
+
 ## Fixed
+
 - Modified Input parameters to be similar to other AzureBlobStorage Tasks.
 
 ## [1.1.0] - 2022-11-28
+
 ### Added
+
 - OAuth autentication
+
 ### Modified
+
 - Unnecessary packages removed:
-	- Azure.Core 1.20.0
-	- MimeMapping 1.0.1.37
-	- Microsoft.CSharp 4.7.0
-	- System.ComponentModel.Annotations 4.7.0
-- Task's result modified: Result.Success added. 
+  - Azure.Core 1.20.0
+  - MimeMapping 1.0.1.37
+  - Microsoft.CSharp 4.7.0
+  - System.ComponentModel.Annotations 4.7.0
+- Task's result modified: Result.Success added.
 - Documentation updated.
 
 ## [1.0.0] - 2022-11-04
+
 ### Added
+
 - Initial implementation of Frends.AzureBlobStorage.CreateContainer.